+++ toolbox/postBuild.js
@@ -1,4 +1,4 @@
-import fs from 'fs';
+import fs, { cpSync } from 'fs';
 import os from 'os';
 import { globSync } from 'glob';
 
@@ -9,6 +9,14 @@ postBuild();
 
 async function postBuild() {
   generateNewIndexDtsFile();
+
+  copyReadmeFile();
+
+  copyAndManipulatePackageJsonFile();
+
+  copyChangesetDirectory();
+
+  copyNpmIgnore();
 
   console.log('DONE !!!');
 }
@@ -37,4 +45,45 @@ function generateNewIndexDtsFile() {
   fs.writeFileSync(`${outFolderName}/index.d.ts`, updatedIndexDts);
 
   console.log('-- Generated new index.d.ts file successfully!');
+}
+
+/**
+ * @description
+ * Copy README file as-is to the output folder.
+ */
+function copyReadmeFile() {
+  console.log('- Step 3: copy the README.md file');
+  const readStreamReadmeMd = fs.createReadStream('./README.md');
+  const writeStreamReadmeMd = fs.createWriteStream(`./${outFolderName}/README.md`);
+  readStreamReadmeMd.pipe(writeStreamReadmeMd);
+}
+
+function copyAndManipulatePackageJsonFile() {
+  console.log('- Step 4: copy & manipulate the package.json file');
+  // Step 1: get the original package.json file
+  const packageJson = JSON.parse(fs.readFileSync('./package.json').toString());
+
+  // Step 2: Remove all scripts
+  delete packageJson.scripts;
+  console.log('-- deleted `scripts` key');
+
+  // Step 3: Change from private to public
+  delete packageJson.private;
+  packageJson.publishConfig.access = 'public';
+  console.log('-- changed from private to public');
+  console.log('-- changed publishConfig access to public');
+
+  // Step 4: create new package.json file in the output folder
+  fs.writeFileSync(`./${outFolderName}/package.json`, JSON.stringify(packageJson));
+  console.log('-- package.json file written successfully!');
+}
+
+function copyNpmIgnore() {
+  console.log('- Step 7: copy the .npmignore file');
+  cpSync('.npmignore', `${outFolderName}/.npmignore`);
+}
+
+function copyChangesetDirectory() {
+  console.log('- Step 5: copy the .changeset directory');
+  cpSync('.changeset', `${outFolderName}/.changeset`, { recursive: true });
 }